split Builder into Artist and MusicData
