+++ src/main/java/entities/GuestUser.java
@@ -1,6 +1,7 @@
 package entities;
 
 import java.util.List;
+import java.util.Scanner;
 
 public class GuestUser extends User {
     public GuestUser(){
@@ -26,7 +27,7 @@ public class GuestUser extends User {
 
     public void logInUser() throws Exception {
         UserData u = new UserData();
-        // ask for email and password
+        // ask fxor email and password
         Scanner email = new Scanner(System.in);
         System.out.println("Enter your email: ");
         Scanner password = new Scanner(System.in);
